refactor(sandbox): share E2B rclone setup
